requirements.txt
gradio>=4.44.0
google-genai>=1.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
